Update Utils, add class related to Firebase (Push)
