+++ app/src/main/AndroidManifest.xml
@@ -74,6 +74,20 @@
             android:windowSoftInputMode="adjustNothing" />
         <activity android:name=".post.PostDetailActivity" />
 
+        <service
+            android:name=".utils.firebase.MyFirebaseMessagingService">
+            <intent-filter>
+                <action android:name="com.google.firebase.MESSAGING_EVENT"/>
+            </intent-filter>
+        </service>
+
+        <service
+            android:name=".utils.firebase.MyFirebaseInstanceIdService">
+            <intent-filter>
+                <action android:name="com.google.firebase.INSTANCE_ID_EVENT"/>
+            </intent-filter>
+        </service>
+
         <!-- <activity -->
         <!-- android:name=".auth.ActivityMissPW" -->
         <!-- android:parentActivityName=".auth.LoginActivity"> -->
