+++ app/src/main/java/com/konkuk/dna/auth/SignupAgreeActivity.java
@@ -9,7 +9,7 @@ import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.CompoundButton;
 
-import com.konkuk.dna.helpers.BaseActivity;
+import com.konkuk.dna.utils.helpers.BaseActivity;
 import com.konkuk.dna.R;
 
 public class SignupAgreeActivity extends BaseActivity {
