+++ app/src/main/java/com/konkuk/dna/friend/fragments/NotifyFragment.java
@@ -11,9 +11,8 @@ import android.widget.ListView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
-import com.konkuk.dna.helpers.AnimHelpers;
-import com.konkuk.dna.helpers.BaseFragment;
-import com.konkuk.dna.helpers.AnimHelpers;
+import com.konkuk.dna.utils.helpers.AnimHelpers;
+import com.konkuk.dna.utils.helpers.BaseFragment;
 import com.konkuk.dna.R;
 import com.konkuk.dna.friend.manage.Request;
 import com.konkuk.dna.friend.manage.RequestListAdapter;
