+++ app/src/main/java/com/konkuk/dna/map/MapFragment.java
@@ -11,7 +11,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
-import com.konkuk.dna.helpers.GPSTracker;
+import com.konkuk.dna.utils.helpers.GPSTracker;
 import com.konkuk.dna.R;
 import com.konkuk.dna.post.Post;
 import com.konkuk.dna.post.PostDetailActivity;
