+++ app/src/main/java/com/konkuk/dna/utils/HttpReqRes.java
@@ -1,4 +1,4 @@
-package com.konkuk.dna.Utils;
+package com.konkuk.dna.utils;
 
 import android.util.Log;
