+++ app/src/main/java/com/konkuk/dna/utils/ServerURL.java
@@ -1,4 +1,4 @@
-package com.konkuk.dna.Utils;
+package com.konkuk.dna.utils;
 
 public class ServerURL {
