+++ app/src/main/java/com/konkuk/dna/utils/firebase/MyFirebaseMessagingService.java
@@ -1,0 +1,50 @@
+package com.konkuk.dna.utils.firebase;
+
+import android.app.NotificationManager;
+import android.app.PendingIntent;
+import android.content.Context;
+import android.content.Intent;
+import android.graphics.Color;
+import android.media.RingtoneManager;
+import android.net.Uri;
+import android.support.v4.app.NotificationCompat;
+
+import com.google.firebase.messaging.FirebaseMessagingService;
+import com.google.firebase.messaging.RemoteMessage;
+import com.konkuk.dna.MainActivity;
+import com.konkuk.dna.R;
+
+import java.util.Map;
+
+public class MyFirebaseMessagingService extends FirebaseMessagingService {
+    private static final String TAG = "FirebaseMsgService";
+
+    // 푸시 메시지를 받았을 때 실행된다.
+    @Override
+    public void onMessageReceived(RemoteMessage remoteMessage) {
+        Map<String, String> pushDataMap = remoteMessage.getData();
+        sendNotification(pushDataMap);
+    }
+
+    // 받은 푸시 메시지를 알림으로 표현한다.
+    private void sendNotification(Map<String, String> dataMap) {
+        Intent intent = new Intent(this, MainActivity.class);
+        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+
+        PendingIntent contentIntent = PendingIntent.getActivity(this, 0, intent, PendingIntent.FLAG_ONE_SHOT);
+        Uri defaultSoundUri= RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
+
+        NotificationCompat.Builder nBuilder = new NotificationCompat.Builder(this)
+                .setSmallIcon(R.mipmap.ic_launcher)
+                .setContentTitle(dataMap.get("title"))
+                .setContentText(dataMap.get("msg"))
+                .setAutoCancel(true)
+                .setSound(defaultSoundUri)
+                .setVibrate(new long[]{1000, 1000})
+                .setLights(Color.WHITE,1500,1500)
+                .setContentIntent(contentIntent);
+
+        NotificationManager nManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+        nManager.notify(0 /* ID of notification */, nBuilder.build());
+    }
+}
