+++ app/src/main/java/com/konkuk/dna/utils/helpers/GPSTracker.java
@@ -1,4 +1,4 @@
-package com.konkuk.dna.helpers;
+package com.konkuk.dna.utils.helpers;
 
 import android.annotation.SuppressLint;
 import android.app.AlertDialog;
