+++ app/src/main/res/layout/activity_post_form.xml
@@ -124,6 +124,10 @@
                     <TextView
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
+                        android:layout_gravity="center_vertical"
+                        android:gravity="center_vertical"
+                        android:textSize="13dp"
+                        android:textColor="@color/gray"
                         android:text="나만 보기"
                         android:tag="nseb" />
